ref: add ZoneBroadcast to base class
