+++ src/Common/Utilities/DataManipulation.cs
@@ -12,4 +12,21 @@ public static class DataManipulation
             .Select(x => Convert.ToByte(hex.Substring(x, 2), 16))
             .ToArray();
     }
+    
+    public static byte[] SpacedHexStringToBytes(string str)
+    {
+        str = str.Replace(" ", "");
+        if (str.Length % 2 != 0) throw new Exception("Hex string must have even number of characters");
+
+        // Convert each pair of characters to a byte and add to the output array
+        var ret = new byte[str.Length / 2];
+        for (int i = 0; i < str.Length; i += 2)
+        {
+            var byteString = str.Substring(i, 2);
+            var b = Convert.ToByte(byteString, 16);
+            ret[i / 2] = b;
+        }
+
+        return ret;
+    }
 }
